+++ src/test/java/com/woowacourse/coordinate/domain/LineTest.java
@@ -1,9 +1,6 @@
-package com.woowacourse.coordinate;
+package com.woowacourse.coordinate.domain;
 
 import java.util.Arrays;
-
-import com.woowacourse.coordinate.domain.Line;
-import com.woowacourse.coordinate.domain.Point;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
