+++ src/test/java/com/woowacourse/coordinate/domain/PointTest.java
@@ -1,8 +1,6 @@
-package com.woowacourse.coordinate;
+package com.woowacourse.coordinate.domain;
 
-import com.woowacourse.coordinate.domain.Point;
 import org.junit.jupiter.api.Test;
-
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.offset;
 import static org.junit.jupiter.api.Assertions.assertThrows;
